17908: Add default devcontainer definition (#29)
Initial attempts to use a badge in the README did not succeed. It
offered to use default dev container images (i.e. jnode, etc.), but not
the Howso Dev Containers. Hopefully, providing a default/Main
devcontainer.json file will allow it to automatically select the dev
container image specified in that file.